+++ DataLayer/Product.cs
@@ -6,13 +6,13 @@ namespace DataLayer
 {
     public class Product : IEquatable<Product>
     {
-        public Product(int price, string name)
+        public Product(double price, string name)
         {
             Price = price;
             Name = name;
         }
 
-        public int Price;
+        public double Price;
         public string Name;
         public string Id;
